Move into lib directory
